Base Version 2
The beginning of a few more functions were added. The new print menu (the restaurants menu) is functional whereas the others are just placeholders.

The programs menu is updated and now fully functional. The operator is now able to add meals to the customers order with this function mostly finished.

Evaluate input function is finished and operational.