Base Version 3
Added functionality means that the of the system is now able to print a customers order, confirm a customers order, and get the total cost of the customers order. This functionality has come through the creation of multiple functions.

The functions that have been added that create this functionality are: print_order, get_price, and confirm order.

Minor bug fixes regarding the programs input validation and some of its functionality. One known bug still to be fixed.