Base Version 4
Added functionality includes the ability to edit the customers order and to edit a meal by removing an ingredient.

The functions that have been added that create this functionality are, print meals, edit order, and edit meal.

Other notable changes include the revamped version of the add meal function, a change in how the menu function determines what it should output, and some improvements made to the print order function.

Minor bug fixes.