Oakwood Kitchen Version 1
Added functionality includes the ability to cancel an order, the ability to  return to the menu from multiple different states, and the ability to select if a customer is going have the food to eat in or takeaway (the program then acting accordingly after).

This functionality comes from the creation of 3 new functions. These are the cancel order function, the return to menu function, and the order type function.

Other notable changes are:
- Reasonably significant format and output changes.
- Major changes to how the code is formatted in order to match conventions.
- The removal of placeholders for the meals, ingredients, and other small things.